separate spsc module
